chatGPT codes example
In this example, the frames are captured from a camera using the cap.read() function, which returns a Boolean indicating whether the frame was successfully read and the frame itself. The frames are then written to the output video file using the out.write() function. The fourcc variable defines the codec to be used for the video, and the (640, 480) tuple defines the resolution of the video. The 20.0 parameter in the VideoWriter function specifies the frames per second (FPS) of the output video.

It is worth noting that the video will be saved in the current working directory, and you can change it to the path that you want to save the video in. Also, you can change the format of the video (avi, mp4, etc.) by changing the extension of the output file.